fix(techdebt): consolidate groupBy duplicates and rollout delta computation
Extract shared groupBy<T,K> to utils/collections.ts, replace 2 local
copies. Reuse buildDeltas() inside buildDecision() in rollout.ts.
